Implement Pattern Tests

+++ SecurityApp/app/src/main/java/com/example/SecurityApp/Attempt.java
@@ -3,19 +3,19 @@ package com.example.SecurityApp;
 import java.util.ArrayList;
 
 public class Attempt {
-    private int attemptTime;
+    private double attemptTime;
     private String lockType;
     private boolean unlockSuccess;
     private String unlockPattern;
 
-    public Attempt(int attemptTime, String lockType, boolean unlockSuccess, String unlockPattern) {
+    public Attempt(double attemptTime, String lockType, boolean unlockSuccess, String unlockPattern) {
         this.attemptTime = attemptTime;
         this.lockType = lockType;
         this.unlockSuccess = unlockSuccess;
         this.unlockPattern = unlockPattern;
     }
 
-    public int getAttemptTime() {
+    public double getAttemptTime() {
         return attemptTime;
     }
